moves logic responsible for handling shortcut event emitting in App.vue component
